requirements.txt
fastapi
uvicorn[standard]
google-cloud-aiplatform
vertexai
requests
python-dotenv
google-cloud-secret-manager